Add files via upload
Learning to work my way through the concept of Stacks using OOP Object Oriented Programming. This first creation of a stack utilizes Arrays and the LIFO philosophy. With simply commands like .size() an equivalent of .length() and isEmpty() to check whether the Stack is empty or not. .push() adds elements to our stack and .pop() removes elements from our stack. The last item added is the first item to be removed which is what LIFO signifies. Last In, First Out